+++ Core/Source/GUIStyles.cs
@@ -36,7 +36,6 @@ public static class GUIStyles
     internal static class Table
     {
         internal const float RowHeight = Text.LineHeight + TableCell.PadVer * 2f;
-        internal const float ExpandedRowHeight = Text.LineHeight * MultiValueDisplay.MaxLines + TableCell.PadVer * 2f;
         internal const float HeadersRowHeight = RowHeight;
         internal static readonly Color ColumnSeparatorLineColor = new(1f, 1f, 1f, 0.05f);
         internal static Color FixedPartSeparatorLineColor => HeadersRowBGColor;